@@ -1,6 +1,6 @@
 package fr.jorisfavier.movietest.models.dto
 
-class ChangedMovieDTO(val id: Int,
-                      val adult: Boolean) {
-
-}
+data class ChangedMovieDTO(
+    val id: Int,
+    val adult: Boolean,
+)
