@@ -1,3 +1,3 @@
 package fr.jorisfavier.movietest.models.dto
 
-class GenreDTO(val id: Int, val name: String)
+data class GenreDTO(val id: Int, val name: String)
