@@ -1,3 +1,3 @@
 package fr.jorisfavier.movietest.models.dto
 
-class ResultDTO<E>(val results: List<E>)
+data class ResultDTO<E>(val results: List<E>)
