@@ -3,33 +3,28 @@ package fr.jorisfavier.movietest.ui.moviedetail
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import fr.jorisfavier.movietest.manager.IMovieManager
+import androidx.lifecycle.viewModelScope
 import fr.jorisfavier.movietest.models.Movie
-import fr.jorisfavier.movietest.models.dto.MovieDetailDTO
-import retrofit2.Call
-import retrofit2.Response
+import fr.jorisfavier.movietest.repository.MovieRepository
+import kotlinx.coroutines.launch
 
-class MovieDetailViewModel: ViewModel(){
-    lateinit var movieManager: IMovieManager
+class MovieDetailViewModel(
+    private val movieId: Int,
+    private val movieRepository: MovieRepository,
+) : ViewModel() {
+
     private var _movie = MutableLiveData<Movie>()
+    val movie: LiveData<Movie> = _movie
 
-    fun getMovie(): LiveData<Movie> {
-        return _movie
+    init {
+        loadDetail()
     }
 
-    fun loadDetail(movieId: Int){
-        movieManager.getMovieDetail(movieId, object : retrofit2.Callback<MovieDetailDTO>{
-            override fun onFailure(call: Call<MovieDetailDTO>, t: Throwable) {
-            }
-
-            override fun onResponse(call: Call<MovieDetailDTO>, response: Response<MovieDetailDTO>) {
-                if(response.body() == null){
-                    return
-                }
-                _movie.postValue(Movie.from(response.body()!!))
-            }
-        })
+    private fun loadDetail() {
+        viewModelScope.launch {
+            val detail = movieRepository.getMovieDetail(movieId)
+            _movie.value = detail
+        }
     }
-
 
 }
