@@ -1,32 +1,36 @@
 package fr.jorisfavier.movietest.ui.movielist
 
-import android.content.Intent
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
-import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import fr.jorisfavier.movietest.api.MovieDbService
+import fr.jorisfavier.movietest.databinding.MovieViewholderBinding
 import fr.jorisfavier.movietest.models.Movie
-import fr.jorisfavier.movietest.ui.moviedetail.MovieDetailActivity
-import kotlinx.android.synthetic.main.movie_viewholder.view.*
 
-class MovieViewHolder(movieView: View): RecyclerView.ViewHolder(movieView) {
+class MovieViewHolder(
+    private val onItemClicked: (Int) -> Unit,
+    private val binding: MovieViewholderBinding,
+) : RecyclerView.ViewHolder(binding.root) {
 
-    private var view = itemView
+    private var movie: Movie? = null
+
+    init {
+        binding.root.setOnClickListener {
+            val id = movie?.id ?: return@setOnClickListener
+            onItemClicked(id)
+        }
+    }
 
     fun bind(movie: Movie) {
-        Glide.with(view)
-            .load(MovieDbService.imageBaseUrl+movie.posterPath)
-            .placeholder(ColorDrawable(Color.GRAY))
-            .centerCrop()
-            .into(view.movie_poster)
-        view.movie_name.text = movie.title
-        view.setOnClickListener {
-            val intent = Intent(view.context,MovieDetailActivity::class.java).apply {
-                putExtra(MovieDetailActivity.movieIdKey,movie.id)
-            }
-            view.context.startActivity(intent)
+        this.movie = movie
+        with(binding) {
+            Glide.with(root)
+                .load(MovieDbService.imageBaseUrl + movie.posterPath)
+                .placeholder(ColorDrawable(Color.GRAY))
+                .centerCrop()
+                .into(moviePoster)
+            movieName.text = movie.title
         }
     }
 }
